requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses==0.6
filetype==1.2.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.0.0
pypub3==2.0.6
pyxml3==0.0.4
requests==2.32.3
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.3
